+++ app/Http/Controllers/TUController.php
@@ -42,7 +42,7 @@ class TUController extends Controller
             'id_sekolah' => 'required|integer',
             'gender' => 'required|string|max:255',
         ]);
-
+        
         $tu = TU::create($request->all());
         return response()->json($tu, 201);
     }
@@ -66,20 +66,48 @@ class TUController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, string $id)
-    {
-        $tu = TU::findOrFail($id);
-        $tu->update($request->all());
-        return response()->json($tu, 200);
+    public function update(Request $request, string $nip)
+{
+    // Validasi input data
+
+    $validatedData = $request->validate([
+        'nama' => 'required|string|max:255',
+        'alamat' => 'required|string|max:255',
+    ]);
+    
+    // Cari data Tata Usaha berdasarkan NIP
+    $tu = TU::where('nip', $nip)->first();
+    
+    if (!$tu) {
+        return response()->json(['message' => 'Data not found'], 404);
     }
+
+    // Update data berdasarkan field yang telah divalidasi
+    $tu->update([
+        'nama' => $validatedData['nama'],
+        'alamat' => $validatedData['alamat'],
+    ]);
+
+    return response()->json([
+        'message' => 'Data updated successfully',
+        'data' => $tu
+    ], 200);
+}
+
+
+    
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $nip)
     {
-        $tu = TU::findOrFail($id);
+        $tu = TU::find($nip);
+        if (!$tu) {
+            return response()->json(['message' => 'Data not found'], 404);
+        }
         $tu->delete();
         return response()->json(null, 204);
     }
+
 }
